pyproject.toml
[tool.poetry]
name = "snaptext"
version = "0.1.0"
description = "Simple OCR tool to extract text from images."
authors = ["Your Name <you@example.com>"]
packages = [{ include = "core" }, { include = "cli" }, { include = "server" }]

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0"
pytesseract = "^0.3.10"
pillow = "^10.0"
opencv-python = ">=4.8.0,<4.12"
numpy = ">=1.24.0,<2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-cov = "^6.2.1"
pytest-mock = "^3.14.1"
requests = "^2.32.5"
psutil = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
